requirements.txt
# Web API
fastapi
uvicorn[standard]

# Telegram bot
aiogram

# Background tasks
arq

# Database
peewee
sqlalchemy
alembic
aiosqlite

# Redis
redis

# HTTP clients
httpx

# Scraping
playwright

# Settings/validation
pydantic[email]
pydantic-settings

# Logging
structlog

# Testing
pytest
pytest-asyncio

# Development tools
pre-commit
ruff

# Admin panel
starlette-admin
python-multipart
passlib==1.7.4
python-jose[cryptography]
itsdangerous
